backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
psycopg2-binary
email-validator 